context-rot/requirements.txt
pandas>=2.2.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
tiktoken>=0.9.0
tqdm>=4.65.0
openai>=1.84.0
anthropic>=0.52.0
google-cloud-aiplatform>=1.95.0
python-dotenv>=1.0.0
python-levenshtein>=0.27.1